[REF] Major consolidation of pypmi.datasets
Reduces submodule down to two main files: fetchers.py and loaders.py,
containing all the relevant information for downloading and loading data
available through the PPMI.
