+++ pypmi/datasets/__init__.py
@@ -1,16 +1,15 @@
 # -*- coding: utf-8 -*-
+"""
+Functions and utilities for fetching and loading PPMI datasets
+"""
 
 __all__ = [
-    'get_studydata', 'get_behavior', 'get_biospecimen', 'get_datscan',
-    'get_demographics', 'get_genetics', 'available_studydata',
-    'download_studydata', 'available_genetics', 'download_genetics'
+    'load_studydata', 'load_behavior', 'load_biospecimen', 'load_datscan',
+    'load_demographics', 'load_genetics', 'available_studydata',
+    'fetch_studydata', 'available_genetics', 'fetch_genetics'
 ]
 
-from .all import get_data as get_studydata
-from .behavior import get_data as get_behavior
-from .biospecimen import get_data as get_biospecimen
-from .datscan import get_data as get_datscan
-from .demographics import get_data as get_demographics
-from .genetics import get_data as get_genetics
-from .utils import (available_studydata, download_studydata,
-                    available_genetics, download_genetics)
+from .fetchers import (available_studydata, fetch_studydata,
+                       available_genetics, fetch_genetics)
+from .loaders import (load_behavior, load_biospecimen, load_datscan,
+                      load_demographics, load_genetics, load_studydata)
